+++ autohooks/terminal.py
@@ -14,36 +14,38 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-import curses
+
+import os
 
 from contextlib import contextmanager
 
 from typing import Callable, Generator
 
-from blessings import Terminal as Term
+import colorful as cf
 
 
 class Terminal:
     def __init__(self):
+        self._width = None
         self._indent = 0
-        try:
-            self._term = Term()
-        except curses.error:
-            # handle issues with terminals and force not to style anything
-            # should not be necessary with blessings > 1.7 anymore
-            self._term = Term(force_styling=None)
+
+    def _check_size(self):
+        self._width, _ = os.get_terminal_size()
 
     def _print_end(self, message: str, status: str, color: Callable) -> None:
-        width = self._term.width - 1
         extra = 4  # '[ ', ' ]'
-        with self._term.location():
-            self.print(
-                message,
-                self._term.move_x(width - extra - len(status)),
-                '[',
-                color(status),
-                ']',
+        # python is adding a ' ' between strings if used
+        # in print('foo', 'bar', 'baz') is printed to "foo bar baz"
+        self._check_size()
+        if self._indent > 0:
+            message = ' ' * self._indent + message
+        if self._width > 0:
+            message += ' ' * (
+                int(self._width) - len(message) - extra - len(status)
             )
+        print(
+            message + '[', color(status), ']',
+        )
 
     @contextmanager
     def indent(self, indentation: int = 4) -> Generator:
@@ -57,21 +59,27 @@ class Terminal:
     def add_indent(self, indentation: int = 4) -> None:
         self._indent += indentation
 
+    def reset_indent(self) -> None:
+        self._indent = 0
+
     def print(self, *messages: str) -> None:
-        with self._term.location(x=self._indent):
-            print(*messages)
+        msg = ''
+        if self._indent > 0:
+            msg = ' ' * (self._indent)
+        msg += ' '.join(messages)
+        print(msg)
 
     def ok(self, message: str) -> None:
-        self._print_end(message, 'ok', self._term.green)
+        self._print_end(message, 'ok', cf.green)
 
     def fail(self, message: str) -> None:
-        self._print_end(message, 'fail', self._term.red)
+        self._print_end(message, 'fail', cf.red)
 
     def error(self, message: str) -> None:
-        self._print_end(message, 'error', self._term.red)
+        self._print_end(message, 'error', cf.red)
 
     def warning(self, message: str) -> None:
-        self._print_end(message, 'warning', self._term.yellow)
+        self._print_end(message, 'warning', cf.yellow)
 
     def info(self, message: str) -> None:
-        self._print_end(message, 'info', self._term.cyan)
+        self._print_end(message, 'info', cf.cyan)